patch from benny:
"One of the things I noticed about your implementation is that you use
ThunarVfsPath's and determine the ThunarVfsInfo on-demand. As said on
thunar-dev, this can cause trouble if the info takes long to load (i.e.
it's a symlink to a slow device or a network share, etc.). Therefore,
unless you receive a change notification, the info should also be
determined asynchronously using thunar_vfs_listdir(). The attached patch
changes XfdesktopFileIcon to use ThunarVfsInfo directly instead of
ThunarVfsPath. There was also a bug in there which caused xfdesktop to
crash when the file name contains non-UTF-8 chars, because you used the
path name directly, instead of using ThunarVfsInfo->display_name, which
is garantied to be UTF-8 (while the path name is always in the local
file system encoding).

"I also changed the "Open" action for folders to try
org.xfce.FileManager.Launch() first, prior to falling back to hardcoded
Thunar invocation. This way, things will also work with other file managers.

"And for thunar_vfs_info_execute() you passed the path of the executable
as parameter (-> path_list), which was probably not what you wanted
(yeah, I know I should write better documentation), since this will
actually expand to exec("<executable> <executable>"). The path_list
parameter for _execute() gives the paths that should be passed as
parameters to the command."


(Old svn revision: 19852)