Use Pango native ellipsizing support; highlight the active window's menu item.
For the ellipsis: I noticed that the width at which truncation occurs causes
display oddness where the "remove workspace" text is somewhat longer than the
truncated text. This isn't right.

For the highlight: I noticed a question about why the active window isn't
highlighted (in #xine) and decided to have a look - this is the reason for
the markup usage.

menulist_set_label_flags() is needed by both changes, so it makes sense to
me to lump them together.

Signed-off-by: Darren Salt <linux@youmustbejoking.demon.co.uk>

(bug 2127)



(Old svn revision: 22764)