Redo how the menu system works so i can support multiple menus on the desktop propery.  Instead of the old method, where main.c watched the desktop for mouse clicks, not XfceDesktop does that, and emits two new signals: populate-root-menu and populate-secondary-root-menu.  Handlers of these get passed a GtkMenu, which they should fill with menu items. They can (optionally) check to see if the passed GtkMenu already has menu items in it (put there by another handler) and put menu items in a submenu of the passed item.
This cleaned up interfaces quite a bit and moved code to some more sane
locations.  The semi-downside of this is that now XfceDesktopMenu needs
to generate a GtkMenu tree from the XfceMenu every time the menu has
popped up.  Performance doesn't seem to be hurt too much on my system,
but it might be a problem for some really slow old hardware (but hopefully
not).  The upside is that the full GtkMenu tree isn't kept around in
RAM all the time.

End-user behavior hasn't changed yet, but this will let me properly fix the
context menu for the desktop when file icons are enabled.


(Old svn revision: 26158)