fix menu module crash on .desktop file change or icon theme change (bug 3615)
this also fixes the weird "double menu" problem (bug 4342).  what was
happening was that thunar-vfs was firing off a dir change notification,
and then the menu would start to regenerate.  then it would call into
GtkIconTheme, which then noticed that the icon cache was stale and
needed to be re-read.  so then it would fire GtkIconTheme::changed,
which causes a recursive menu regeneration.  since the menu generation
function isn't reentrant, now we use g_idle_add() to schedule a menu
regen, and we don't schedule one if there's one already scheduled or
in progress.

(Old svn revision: 27912)