drawing routines overhaul
this is mainly a lot of refactoring and moving common code into
functions.  XfdesktopIcon keeps track of a bit more information about
its position and the extents of its pixbuf and text area.  this should
fix most (if not all) drawing glitches, though there are a few hacks in
place that i need to fix.