Make XfdesktopFileIconManager use a GFile. Use CreateFiles from Thunar.
The file icon manager now uses a GFile for its folder instead of a
ThunarVfsPath. There are a number of situations where we still need to
create a ThunarVfsPath for this GFile again, e.g. for monitoring and the
template action in the context menu. But this will soon be changed.

We're now using the org.xfce.FileManager.CreateFiles() method to create
new directories. This allows us to re-use the create dialog from Thunar.