Add helper methods to spawn GAppInfos, similar to Thunar.
The new methods are:
  * xfdesktop_file_utils_app_info_launch()
  * xfdesktop_file_utils_file_is_executable()

The first one launches a GAppInfo and properly sets the working
directory before and after the application is spawned.

The second determines whether a file can be executed. This depends on
the executable bit as well as the content type and the file extension
(.desktop can be executed, .directory can't).

Check for sys/params.h and pwd.h.