Properly open mount point folders in the file manager after mounting.
This is done via an "activated" GQuark that is set when a volume icon is
activated. It is checked (and reset) in the mount finish callback and
used to chain up the activated event to the XfdesktopFileIcon class
where the mount point is checked and opened in the file manager.