Add menu item monitoring.
Reload the garcon menu if the reload-required signal is triggered
and make XfdesktopAppMenuItem use the data inside GarconMenuItem.

I've dropped the cache since it's not function as it should be,
and because the GarconMenu is not destroyed, the menu information
is still presereved in memory. This could be improved tho.

Because the menu items are destroyed to between 2 popups, the
changed-signal in the GarconMenuItem isn't very functional either,
but it is implemented.