Use the right icon style in the desktop settings by default(bug #6798).
When running xfdesktop for the first time, the /desktop-icons/style will
be undefined. This caused the desktop settings dialog to always show
'None' as the icon style even though xfdesktop uses file icons / window
icons by default (depending on whether file icons are built in). So we
now follow that behaviour in the desktop settings dialog.