Let Garcon decide what menu file to choose. Drop UserMenu kiosk option.
Garcon now supports XDG_MENU_PREFIX which we can use to make
xfce-applications.menu the default menu file in all our applications.
This way, xfdesktop no longer needs to decide which file to pick on
its own. The bad news is that this makes the UserMenu kiosk option
nearly impossible. Therefor I decided to drop it; we don't restrict
the application menu in the panel either, so why should we do here?