Use Name field as display name of desktop files (bug #6896).
To implement this in a somewhat clean way, the function
thunar_g_file_query_key_file() was copied over as
xfdesktop_file_utils_query_key_file(). Another method
xfdesktop_file_utils_get_display_name() was added to determine the
correct display name to use for a given file and file info.

XfdesktopRegularFileIcon was altered to include a private variable
display_name in which the cached display name is stored.