Handle volume changes in a timeout (bug #7025).
When the "volume-changed" signal of GVolumeMonitor or the "changed"
signal of GVolume are emitted after a volume has been mounted, then for
some reason the affected GVolume still has no valid GMount. Same happens
when unmounting: the GVolume still has a valid GMount.

This commit works around this by checking for a change from valid to
invalid GMount or from invalid to valid GMount in a timeout that is
called at most 5 times until it gives up. When a change is detected, the
corresponding XfdesktopVolumeIcon is updated.

I hope this is a suitable workaround for most machines/drives. The
underlying issue seems to be in GIO or GVfs. It might be worth finding
the root cause.