Fix crash when cancelling a file download (bug #6876).
Apparently, there are situations where a G_FILE_MONITOR_EVENT_CHANGED
event is triggered for an existing file that has no GFileInfo any more.
This commits removes an assertion that the GFileInfo is valid in those
situations and instead, destroys the file icon when it's not.

Also, this commit switches from G_FILE_MONITOR_EVENT_CHANGES_DONE_HINT
to G_FILE_MONITOR_EVENT_CHANGED although it shouldn't really matter
which one we use.