Stop altering the root window pixmap to avoid X crash (bug #7442).
Altering the back pixmap makes X segfault, sometimes on startup
of xfdesktop, all the time during logout. Not the best solution
to fix it this way, but the background is still shown when
xfdesktop is running, so not a big issue.

(cherry picked from commit bf07cc01b5457067303bdfad93433def548b6172)