Thumbnail desktop icon previews using tumblerd
Adds support for drawing thumbnails from the dbus thumbnail
service. Adds a show-thumbnails xfconf property to toggle showing
thumbnails. Moves the marshal.list into the common folder. Creates
a checkbox option on the xfdesktop-settings app to toggle
thumbnails. Adds a tooltip-size gtk style property to change the
tooltip image size, ranges from 0 (not shown) to 512, with 128
as the default. Added a timer when the icon size spinner changed
value in the settings app to prevent the icons redrawing on every
value change. Removed xfdesktop_icon_mark_extents_dirty so that
size changes to the pixbuf are properly rendered. Bug 4344.