Arrange icons on the desktop
This patch adds code to arrange the desktop icons. First,
it sorts and adds all the special icons such as the trash, home
folder, and any volumes. Next, it sorts and adds all the folder
icons. Finally, it adds the regular file icons. It can be invoked
by  a root menu option or via a command line --arrange so it can
be used as a keyboard binding.
The xfdesktop_icon_view_compare_icons function was written by
Nick Schermer <nick@xfce.org>
For bug 2833.