Fix for moving files instead of copy when src isn't writable
Fixed a bug where a user drags and drops a file from a folder on the
same filesystem as the desktop but the src isn't writable by the user
so they can't delete the src file. The user will still see an error
message if they explicitly try to move the file via the right click
drag and drop menu.