Improve menu icon loading (Bug #8795)
This code checks if the icon theme has the custom icon name and
loads that directly into a gtk_image widget rather than a pixbuf
first, which should provide a small speed improvement when many
items are present in the menu. Should that fail, it will fallback
to checking all other possible locations for the menu item's icon.