Port fix for bug #9892 from the xfce-4.10 branch
With the recent addition of per-workspace wallpapers (commit
88a920697b22b0653f244ede15f8e645ae013139) this bug is much harder to
trigger but can still manifest when an invalid image is selected and
style is set to Tiled. (xfce_backdrop_get_pixbuf() will hang in a for
loop.)

This commit ports the original fix for consistency and correctness
reasons. In addition, it also addresses the following warning which
appears when style is set to None:

  ** (xfdesktop:3236): CRITICAL **: Invalid image style: 0