Fix for invalid single workspace number in settings app
Xfdesktop-settings app now repsonds to workspaces being created
and destroyed because it can impact how the single workspace mode
operates. If the single workspace number is set to a workspace that
is destroyed it reverts back to per workspace wallpapers and toogles
the checkbox so that the user knows what is going on.