Cache wallpapers when they are loaded
This is to make workspace switching faster. It only cached when drawing the
wallpaper because the user may have lots of workspaces configured for whatever
reason and never use all of them, so this would save on the memory overhead.
This patch also improves the backdrop comparison check and simplifies the
unbinding of some settings.