Use an async message queue to load preview images
This patch uses the g_async_queue for message passing between the
main thread and the preview thread of the settings app. Doing this
gets rid of the idle timeout callback and a seperate thread loading
function. It also makes it easy to pop any pending previews when
the app changes what the icon view should display.