xfdesktop-settings is more responsive
Changed to using g_file_enumerate_children_async and enumerating
each file individually in an idle callback function to keep things
responsive while loading large directories of images. This is also
cancellable so that when the user moves the window to a different
monitor or workspace it will respond faster.