Change tooltips in xfdesktop-settings
Since the tooltips were reporting the image size of the thumbnails
and parsing the original file just to get the image dimensions is
slow, this patch changes it to pull the content type and file size
to populate the tooltip with. Also fixed how tooltips in the icon
view popped up so that it follows the mouse when it is sensitive.