More consistent icon placement (Bug #8814)
When the screen size changes, icons are loaded from the icon cache
if the position is available, then an attempt is made to re-locate
the icon at its original position before the change event, if that
is still available, and finally any leftover icons are appended to
any free position. This should help keep the icons where the user
would like them to be.