Fix for CTRL + drag selections (Bug 10275)
When a rubber band selection area shrinks, the icon must be in the
old rubber band area and not in the new one for it to be removed.
Otherwise, when doing a rubber band with the CTRL modifier key any
time the area shrinks it will de-select all the icons that weren't
in the new rubber band area.