Align icon labels and allow them to be multi-line (Bug 8646)
When thumbnails are turned on the icon labels can become unaligned with others
in that row. There's also unused space in the icon's label area that is now
used to display more of the text. Additionally, some effort was made to lower
the number of times the icon redraws itself in response to expose events,
especially during xfdesktop's startup.