Better icon pixbuf caching
With the inclusion of tooltip pixbuf previews the icon pixbuf would
get recreated every time the tooltip pixbuf was generated. So now it
will cache both the normal size icon pixbuf and the tooltip pixbuf.
The pixbufs are stored in the base icon so that all the different
icon types don't have to do the same redundant checks.