Monitor the gvfs directory
Added a file monitor watching the gvfs directory since there's no
other way to be notified of these events. Now when the user adds
an emblem to a file and gvfs commits the change to the db, xfdesktop
will update the icons to reflect this change. There is about a one
minute delay from setting the emblem and gvfs changing the file.