Change how icon positions are saved (Bug 9192)
First a version stamp has been added to the RC file so that it will
hopefully preserve most of the old positions to the new format.
Next, most icons are stored based on the file path rather than the
icon label. This way icons with the same label don't cause an issue.
For volume icons the uuid of the icon is used (when available),
otherwise it will fallback to the label.