Use the GApplication API
Uses the GApplication API to handle process uniqueness, message
passing, and lifecycle management. This eliminates the need to
check if xfdesktop is already running and using X to forward events
to the running instance of xfdesktop making it eaiser to port to
GTK3 in the future (as well as GTKApplication).