Wait for the window manager (Bug 7769)
When xfdesktop starts before the window manager the wallpaper images
and icons won't appear on all the screens. This patch waits for the
window manager to start up or 5 seconds before loading the desktops.
Most of the actual checking if the wm is ready is based on a patch
in xfce4-panel, commit-id cabbdfd1c04caf20a71e0af773accf809135a03d