Add icons in an idle callback
All icons are now added to the desktop in an idle callback. Icons
which had a position on the desktop already or have a cached icon
location are added first and pending icons are added last. This
should make xfdesktop more responsive to user input while icons
are being loaded as well as give the backdrop time to load at the
same time.