Add a chronological backdrop cycle option
When cycling backdrops sequentially with a period of hourly, setting
this option causes xfdesktop to map files in the selected directory
to hours of the day. Fewer images in the directory than hours in the
day are handled by repeating images to scale them over the time range.
That means there will be at most 24 images displayed showing the first
image from midnight to one, the second from one until two o'clock,
etc. If there were 6 images in the directory then every image would
be displayed for 4 hour time intervals before advancing. If there's
more images in the folder than hours then they won't be shown.
The entire point is that the user can have darker backdrops shown at
night or do other fancy time of day image selections (i.e. one image
in the morning, one in the afternoon, one in the evening, etc).