Fix template sub-menu loading code (Bug 10138)
The old code recursively searched the XDG_TEMPLATES_DIR for files
that can be used as templates. This patch changes it so that it
only searches that directory and limits the number of items shown
to the xfconf property max-template-files which is set to 16 to
keep the sub-menu usable. It also only shows files which are not
marked as hidden or backup to eliminate clutter.