Speed up template sub-menu loading code (Bug 7834)
This patch will make it so that xfdesktop won't walk the entire
directory and throw away most of the results when there's more
than max-templates present. Instead it only gathers up until
max-templates then displays those files. This should help when
the XDG_TEMPLATES_DIR has a lot of files in it.