Fix icon moves and deletions
When creating a temporary file and renaming it, sometimes the icon
wouldn't appear on the desktop because the file's info was NULL.
This patch fixes how moves and deletions are handled to fix that.
Additionally we monitor the changes done event rather than just
the change event to prevent flickering icons when large files
are moved or downloaded to the desktop.