Improve how backdrops are choosen
The patch changes it so that xfdesktop keeps the list of possible
images in the same directory as the current file and sorts them
the same way as xfdesktop-settings. This is to help make it more
predicatable when xfdesktop advances the image. It will also monitor
the directory of images for image files being added or removed.