Refresh desktop on style changes
Xfdestkop's icon view doesn't update the icon labels when xfsettingsd
starts up. To work around that, refresh the desktop on a style change
event. Additionally gtk likes to send up to 10 style change events at
once, so this also adds a 1 second delay to ensure it's the last change
event that we respond to.