Clean up some user strings
1) Some tooltips end with a period, some do not. This has been
unified.

2) "You are using more than one display, move this dialog to the
display you want to edit the settings for." This takes a lot of
space and brings along a bit of redundancy. Changed to "Move this
dialog to the display you want to edit the settings for."

3) "Other items" on the third tab, below removable devices: "devices"
is a better term than "items". Changed.

Thanks to Harald Judt for pointing these issues out.