Better migration from previous version
This code makes xfdesktop set whatever settings it finds to xfconf.
This way it isn't always trying to migrate if the user never defines
anything. There was also a couple bugs in the migration code that
have been fixed. Additionally, when gtk sends a style set we only
recreate the icon view (to update the labels) rather than refrsh
the wallpaper as well. This should make xfdesktop startup a little
faster. The style set workaround will go away in GTK3 when we get
there.