Keep the .desktop.in file
The .desktop.in file is generated by the .desktop.in.in file,
however translations are applied against the desktop.in so we'll
keep it in tree so that translations continue to work. The only
thing that will change in the desktop.in file is the version string
during stable/development changes (i.e. from 4.11 to 4.12).