Improve style changes for the desktop
This patch changes how style-set events are handled. Instead of
re-creating the icon view to deal with XSettings manager changes,
we now reset all the styles on the icon view and let it re-set them
to what it needs to. This stops the icon view from destroying and
recreating all the icons which should make startup a little faster.