Reset the icon_view style on style change (Bug 10605)
Since xfdesktop now properly responds to xfsettingsd changes we
need to handle the race condition at launch. This patch changes
it so that the style change for the icon view happens in a low
priority callback. This ensures that the style gets reset only
once if there's a batch of changes to process and will happen
after xfdesktop has loaded. This is different from the previous
callback patch in that the entire icon view doesn't get recreated.