Filter out changed events for special icons (Bug 10755)
Special file icons are the trash, home, and filesystem icons. The
only events that would require a refesh would be an add/delete
event for the trash can. This patch filters out all the change,
attribute change, and changes done hints to prevent the icons
from being reloaded due to those events.