Fix for monitors not getting updated
When multiple monitors are connected there are cases where the
backgrounds don't get drawn. This patch makes it so that the
backdrop is always drawn on a workspace change. Because of that
change it makes sense to always cache the backdrop image since it
would have to reload the image/canvas often.