Add runtime debug support
This patch adds the ability to enable and disable debug messages
for xfdesktop while it is running. It also converts most of the
DBG messages to XF_DEBUG ones except where it would cause a lot
of noise. Additionally, when built with --enable-debug=full the
XF_DEBUG messages will be on by default. The point of this patch
is to allow the users to easily provide more details in their bug
reports.