Add a --next option to force wallpaper changes (Bug 10754)
Several users have external scripts telling xfdesktop when to
change the wallpaper. This patch adds a --next option so that
users can advance the wallpaper on the current workspace when
they want with those scripts, rather than being limited to only
the time based events that are built into xfdesktop.