settings: Fix background color loading
The issue was launching xfdesktop-settings changed the
wallpaper background color, either to black or #7f000 or
some other color.
There's a bit of a mixup in the code, there's color 2s where
there should be color 1s. With these corrected the colors
should properly be loaded and old versions should be migrated.
Bug reported to launchpad but never sent up to xfce's bugzilla:
https://bugs.launchpad.net/ubuntu/+source/xfdesktop4/+bug/1313728