Remove the drag highlight code (Bug 10450)
When dragging icons on the desktop there was a box showing which
cell the icon was going to be placed in. The issue is that the
bounding box doesn't always get cleared and ends up causing
different artifacts on different systems. For now, remove this
feature until it can be correctly implemented.