Use gdk's get static image for animation files
This way it will return a reasonable thing to display as a static
unanimated image, which might be the first frame, or something
more sophisticated.

And since it's only a little bit more work, add a hidden property
to display animations as a wallpaper, the xfconf property is
/backdrop/screen#/MonitorName/Workspace#/backdrop-do-animations
such as:
/backdrop/screen0/monitoreDP1/workspace1/backdrop-do-animations
it's a boolean value and setting it to TRUE turns it on, default
is obviously off.

Signed-off-by: Eric Koegel <eric.koegel@gmail.com>