Fix next background (!16)
Since libxfce4ui was ported to gtk3, "Change the background" is checked
for only the first monitor, if enabled, "Next Background" is available
to all monitors. When "Next Background" is activated, all monitors with
"Change the background" enabled will have their wallpapers cycled.

With this commit, "Next background" is only available to monitors with
"Change the background" enabled and activating it will just update the
wallpaper of the current monitor.