Fix X11 multiscreen support
Not to be confused with multimonitor support, there are still some
setups out there (often people with multiple GPUs) where X exposes
multiple monitors as multiple X11 screens.  This concept is technically
deprecated in GTK3, but still works if you ask for the default GdkScreen
on the display, assuming xfdesktop was started with the appropriate
DISPLAY setting that points to the desired screen.