Refactor how root menu population works
Previously it was abusing gobject's signal system to attempt to layer
menu population from the icon view and from the root desktop.  This was
confusing and awkward, and from a time when I overengineered things
(even more than I do now), for a purpose that was not ever really
necessary.

Now XfceDesktop just knows whether or not to populate from the icon
view, depending on which menu was desired, and then layers on the proper
root menu contents.

This also keeps the code that actually pops up the menu in one place,
simplifying things greatly.