Possibly fix issues finding icon locations in the rc file
The code for determining what to use for the rc file group name for each
icon is a little confusing.  There's the icon identifier, and the icon
label.  At some point using the icon identifier was preferred over the
label, but I think in some cases one of those two might get written to
the file, while the *other* ends up being used to read it back.

Frankly, I wish this had been done in a way where we just migrate old
data to the new method, and then mark the rc file as having been
migrated, but hopefully this will give us the best chance of at least
reading *something* out of the rc file.