Call file manager launch API asynchronously
When launching a desktop icon, the launch API might block for a while
(if the file manager puts up a dialog, for example).  This will block
xfdesktop's UI thread, and undesirable things (like tooltips not hiding)
will occur.

Running asynchronously fixes that, however there's a possibly annoying
side-effect: in some of these cases, it's actually probably preferable
that these sorts of dialogs in effect act as modals and block user input
on the desktop window.  Now that behavior is gone.  Ideally I should
de-sensitize the icon view while this dbus call is running, but how
things are currently set up makes that not trivial to do.