Better context menu population with multiple selected icons
If all selected icons are similar -- that is, they have the same list of
applications that can open them -- act as though a single file is
selected, and populate the context menu with all the applications that
can open them.

If an application is selected, pass all selected files to a single
execution rather than one execution per selected file.