Make background previews less blurry when UI scale factor != 1
This is a bit of a hack, because GtkIconView doesn't appear to allow us
to handle different scale factors properly when providing GdkPixbuf
instances.

GtkIconView also requires the pixbuf column to be a GdkPixbuf, but to
handle the scale factor correctly I believe we need to use a
cairo_surface_t instead.  So we give GtkIconView its GdkPixbuf column,
but also give it a cairo_surface_t column, and then overwrite the
settings on its GtkCellRendererPixbuf to use the surface instead of the
pixbuf.

Hopefully this doesn't break in weird ways with new GTK versions.