xfdesktop-icon-view: let XfconfChannel be passed in instead of new one
Since we already have a valid XfconfChannel object in the desktop object,
we don't need to allocate a new one inside the icon view. Even more: since
the icon view is just a sub component of the desktop, the desktop should
be in control, which channel the icon view is looking at.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>