Improve behavior when changing folders in settings dialog
If the background cycling option is enabled, and the user selects a new
folder, nothing will be selected in the icon view, but sometimes users
will think that is enough to change the folder where xfdesktop will find
backgrounds to cylce through, and close the window.  Unfortunately, that
doesn't actually change the folder; only selecting a background in the
new folder will do that.

Selecting one automatically (when background cycling is enabled) is
undesirable because users will sometimes navigate around their folder
structure to explore what's available, and will sometimes decide not to
change the background.

This commit adds an "Apply" button next to the folder selection button
when the folder is changed and the background cycling option is enabled.
The Apply button is hidden after it is activated (which selects the
first background in the icon view).

Other options considered:

* Add a banner on the top warning the user that the folder is not
  actually selected.  But we already have a banner there in
  multi-monitor setups, and adding extra text for the user to read
  (especially for users who *don't* forget about this issue) is
  annoying.
* Add a confirmation dialog when closing the settings window if no
  background is selected.  Two issues here: 1) we cannot actually do
  this when the settings dialog is embedded in the Settings Manager, and
  2) this would be an annoyance for people who navigate around and then
     change their mind and want to close the dialog without changing
     anything.
* Add an extra "Select Folder" button to the dialog's action area that
  selects the first item in the folder and then closes the dialog (in
  addition to the "Close" button which would discard changes).  I like
  this option, but unfortunately this isn't possible when the settings
  dialog is embedded in the Settings Manager.