Place new files in the location they were dropped
When files are dropped onto the desktop from an external location, the
move, copy, or link happens asynchronously.  By the time the
GFileMonitor sees the new files and adds them to the view, the location
of the drop is long forgotten.  To fix this, choose positions for the
new files (based on the drop location) when we receive the drop data,
and store these new positions for a while (60 seconds, for now) until
the GFileMonitor sees the icons created.

Closes #14