Move file icons settings to its own section
This allows us to easily make the file icons settinsg insensitive when
icons are disabled, or minimized apps icons is set.  This also makes it
clear that those settings only apply to file icons, and not other icon
styles.

I've also moved the "Show 'Delete' option in file context menu" checkbox
out of the desktop menu settings and into this new frame on the icons
tab, as it's specific to file icons.