Make the icon view not know about workarea
Instead, the icon view manager will track workarea, and will position
the icon view inside its parent (using GtkFixed) based on the workarea.
This way the icon view just doesn't need to deal with it at all.

I'm not 100% happy with how this worked out; it seems things are
sensitive to the order that I do things, which I don't like.

Another annoying bit is that GtkFixed sets the GdkWindow background as a
part of a workaround for a silly behavior that some app cares about, so
we can no longer set a background pattern on XfceDesktop's GdkWindow and
let the X server handle drawing.  This isn't terrible, I guess, but will
probably result in more data being sent between xfdesktop and the X
server occasionally.