Fix apps menu not popping up when menu icons disabled
When disabling menu icons, it causes the menu to stop loading and
reload.  Unfortunately, this changes the behavior of gtk_widget_show(),
which usually waits for the menu's initial load to complete before
returning.  Since the no-icons reload is not the initial load, it
doesn't wait.

This fix creates the menu by *first* setting the show-menu-icons
property before setting the GarconMenu, the latter of which is what
triggers the initial menu load.  This way, gtk_widget_show() will always
wait for the menu to be populated, regardless of the icons setting.