Ditch GtkCellLayout impl for XfdesktopIconView
There's really no reason to be that generic, and trying to insert other
cells from outside the icon view would just cause weird rendering
issues, so get rid of it.

This also fixes icon/text positioning when the icon is not square, and
the width is greater than the height.