main: Fix control (always) reach end of non void function
When configured with --disable-wayland, the ENABLE_WAYLAND will be
dropped, the "return matches" clause will be considered the else
statement of the "if (xfw_windowing_get() == _X11)", which is always
false in this scenario.  Thus, this function always reach end of
function without any return values.

While we're at it, fix the comment for the ENABLE_WAYLAND leg.