Clean up autogenerated source file build rules
There's really no reason to put these behind MAINTAINER_MODE, as the
programs that generate the code are included with the librar dev
package, which is of course required anyway to build xfdesktop.

Additionally, let's check for these programs properly at configure time,
and bail out early if they aren't present.  This also makes it possible
for the builder to override the locations of these binaries if they need
to.