Update for new libxfce4windowing workspace API
This also greatly simplifies how workspaces are tracked and represented
in both the backdrop code and window icon code.  Previously the list of
workspaces was an array indexed by the workspace number, but now it's a
hash table with the key as the XfwWorkspace and the value as the
xfdesktop-specific workspace data.

This removes a lot of weird hacks and makes things overall more reliable
and simpler.

Closes #283.