Refactor the GApplication stuff
I think whoever implemented this didn't entirely understand how
GApplication's option parsing and handling (local vs. remote) is
supposed to work.  So that's fixed now.

I also decided to use GtkApplication instead of GApplication, since it
calls gtk_init() for us and sets up (and un-sets-up on shutdown) other
gtk-related things, some of which are useful for an app like xfdesktop,
some not.

So now we don't call gtk_main() anymore, because GApplication iterates
the main loop under the hood.  This fixes some oddities with how things
were working, including strange behavior on Wayland where running
`xfdesktop --quit` when it wasn't running at all would... cause it to
start.

Still, GApplication is a little hard to work with: there's no good way
to handle the case where xfdesktop isn't running, but it's called with a
remote command like --next or --quit.  So I'm working around that by
checking for remote options in the local option handler, and quitting in
xfdesktop_application_startup().  Not great, but it works.

Closes #288.