Be more precise in what gets redrawn
There are far too many instances of gtk_widget_queue_draw() all over, I
suspect to paper over redraw issues.  Removing all of these (and
replacing them with individual or all icon invalidation) should be fine
to do.  There's still the possibility that some areas won't get
invalidated when they need to be, but they should be fixed when found.
The only time we should `_queue_draw()` is when the backdrop changes.

Additionally, GtkWidget unconditionally queues a redraw in response to
focus-in and focus-out, so now we're preventing that from happening from
the XfdesktopIconView code (which handles much more targeted redrawing).

Closes #284.