Compute icon invalidation rect more generously
Previously we'd invalidate the entire slot only if the item was under a
DnD operation.  Now we unconditionally invalidate the union of the slot
rect and the item's calculated extents, if any.

This might cause a *little* more drawing than necessary in some cases,
but fixes some intermittent drawing glitches when moving icons around.